requirements.txt
sklearn-rvm
streamlit
pandas
numpy
plotly
pillow